supabase/functions/document-extract-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23
mammoth==1.6.0
beautifulsoup4==4.12.2
lxml==4.9.3
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
textract==1.6.5
ebooklib==0.18
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.4
aiofiles==0.24.0